Liveness testing module (#93)
Move PhantomIsLive in registration.go to a independent module so that the detection of the liveness of IP address can be re-used in other context. Added cached liveness with zmap as a way to seed cache. Cache is managed by an independent struct that can be configured using the `cache_expiration_time` field in the application config. The LivenessTester interface is instantiated by the registration manager and integrated with the application to make sure we aren't over scanning the phantom addresses.