update station application to use go-redis/resis/v8 (#108)
This PR updates the application to go-redis/redis/v8 and the applies the proper changes to the way we use the API. This should be compatible with the redis default server used in typical station installs. In support of this change the Docker build process no longer needs to checkout a custom version of the go-redis/redis repo and now. This lead to some modification to the docker workflow that fixes a bug when building using local modifications to source code. Also this PR revamps the CI/CD pipeline to validate tests, linting, and formatting for both rust and golang code.