Enhancements for dtls.SCTPConn (#253)
dtls.SCTPConn is the net.Conn returned from gotapdance.Dial for the Conjure
protocol with the DTLS transport. Callers may expect that net.Conn to behave
as a TCP-like byte stream, and not an SCTP message stream, and these changes
accomodate that. These changes should also be appropriate for use of
dtls.SCTPConn on the station side.

- Skip 0-byte writes
- Buffer reads, allowing the caller to read smaller amounts of bytes without
  hitting the "short buffer" error
- Add write flow control to avoid possibly unbounded outbound packet buffering